refactor: organize code
