+++ packages/tiny-store/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./core";
+export * from "./local-storage";
